Fix ordering, and add some error checking.
Previously, the ord function would not take the project name into account.
This would mean that the sorting would not work, and therefore finding
JD numbers did not always work.  Now all of this is fixed.

Also, I add some error handling, so that if you have an area like
`11-20_label`, it catches it and errors, instead of silently failing.